backend/requirements.txt

firebase-admin
python-dotenv
boto3
celery
redis
